Split command line tests into smaller modules down from one 2K line long monster. Also add a command to reinject the current config into an engine (useful for picking up out-of-band changes like cache sizes).
